+++ lib/edgedb/types/set.ex
@@ -7,7 +7,7 @@ defmodule EdgeDB.Set do
 
   ```iex
   iex(1)> {:ok, client} = EdgeDB.start_link()
-  iex(2)> %EdgeDB.Set{} =
+  iex(2)> set =
   ...(2)>  EdgeDB.query!(client, "\"\"
   ...(2)>   select schema::ObjectType{
   ...(2)>     name
@@ -15,32 +15,33 @@ defmodule EdgeDB.Set do
   ...(2)>   filter .name IN {'std::BaseObject', 'std::Object', 'std::FreeObject'}
   ...(2)>   order by .name
   ...(2)>  \"\"")
+  iex(3)> set
   #EdgeDB.Set<{#EdgeDB.Object<name := "std::BaseObject">, #EdgeDB.Object<name := "std::FreeObject">, #EdgeDB.Object<name := "std::Object">}>
   ```
   """
 
-  defstruct [
-    :__items__
-  ]
+  defstruct items: []
 
   @typedoc """
   A representation of an immutable set of values returned by a query.
   """
-  @opaque t() :: %__MODULE__{}
+  @opaque t() :: %__MODULE__{
+            items: list()
+          }
 
   @doc """
   Check if set is empty.
 
   ```iex
   iex(1)> {:ok, client} = EdgeDB.start_link()
-  iex(2)> %EdgeDB.Set{} = set = EdgeDB.query!(client, "select Ticket")
+  iex(2)> set = EdgeDB.query!(client, "select Ticket")
   iex(3)> EdgeDB.Set.empty?(set)
   true
   ```
   """
   @spec empty?(t()) :: boolean()
 
-  def empty?(%__MODULE__{__items__: []}) do
+  def empty?(%__MODULE__{items: []}) do
     true
   end
 
@@ -51,12 +52,12 @@ end
 
 defimpl Enumerable, for: EdgeDB.Set do
   @impl Enumerable
-  def count(%EdgeDB.Set{__items__: items}) do
+  def count(%EdgeDB.Set{items: items}) do
     {:ok, length(items)}
   end
 
   @impl Enumerable
-  def member?(%EdgeDB.Set{__items__: []}, _element) do
+  def member?(%EdgeDB.Set{items: []}, _element) do
     {:ok, false}
   end
 
@@ -66,7 +67,7 @@ defimpl Enumerable, for: EdgeDB.Set do
   end
 
   @impl Enumerable
-  def slice(%EdgeDB.Set{__items__: []}) do
+  def slice(%EdgeDB.Set{items: []}) do
     {:ok, 0, fn _start, _amount, _step -> [] end}
   end
 
@@ -86,13 +87,13 @@ defimpl Enumerable, for: EdgeDB.Set do
   end
 
   @impl Enumerable
-  def reduce(%EdgeDB.Set{__items__: []}, {:cont, acc}, _fun) do
+  def reduce(%EdgeDB.Set{items: []}, {:cont, acc}, _fun) do
     {:done, acc}
   end
 
   @impl Enumerable
-  def reduce(%EdgeDB.Set{__items__: [item | items]}, {:cont, acc}, fun) do
-    reduce(%EdgeDB.Set{__items__: items}, fun.(item, acc), fun)
+  def reduce(%EdgeDB.Set{items: [item | items]}, {:cont, acc}, fun) do
+    reduce(%EdgeDB.Set{items: items}, fun.(item, acc), fun)
   end
 end