+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rendering hints for query errors from EdgeDB.
 - support for EdgeDB binary protocol `2.0`.
 - new `EdgeDB.MultiRange` type to represent multiranges from `EdgeDB 4.0`.
+- `EdgeDB.Object.id/1` to fetch ID from an `EdgeDB.Object` if it was returned from the query.
+- `EdgeDB.ConfigMemory.new/1` to create a new instance of `t:EdgeDB.ConfigMemory.t/0`.
+
+### Changed
+
+- `t:EdgeDB.Object.t/0` to be `opaque`.
 
 ## [0.6.1] - 2023-07-07